Add extern to C API symbols - update hello_classification_c to include files and catch build error
