+++ samples/c/hello_classification/CMakeLists.txt
@@ -3,5 +3,6 @@
 #
 
 ov_add_sample(NAME hello_classification_c
-              SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/main.c"
+              SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/main.c;${CMAKE_CURRENT_SOURCE_DIR}/infer_result_util.c"
+              HEADERS "${CMAKE_CURRENT_SOURCE_DIR}/infer_result_util.h"
               DEPENDENCIES opencv_c_wrapper)
